Combine existing BigQuery samples into single file
